Move codegen into separate file
